Roll up first non-passing status for scenarios
If a scenario had a pending or missing step, then it seems wrong to report it as "passed". It can mask problems. So this change will set the status of a scenario to the first non-passing status of its steps. So the first pending, missing, etc status will be propagated to the reports.

Fail config flags still work and take precedence. The logic to use the last-non-passing status happens afterward.