Use stable sorting on tables
This is an issue when multiple entries in a list have the same value.
If step A and step B both took 3ms, then a stable sort will guarantee
that every time the list is sorted, A and B will appear in the same
order with respect to another. Some sorting implementations can't make
this guarantee.

By default, tablesorter disables this.

More info: https://en.wikipedia.org/wiki/Sorting_algorithm#Stability