Use counter to ID elements
Previous used hashes, which aren't guaranteed to be unique. So
expand/collapse logic would fail if two elements on the same page had
the same hash. An example would be a failure error message from a
Scenario Outline.

Now using an incrementing counter at the page level to ID elements.