Escape HTML using the velocity engine
It is preferable to escape HTML through the template engine over escaping it at in
the template. This ensures none can be forgotten on accident. The unfortunate side
effect of this is that it no longer possible to  provide html as an argument in a
macro. This leads to some duplication in embeddings.vm

Existing methods to escape html have been removed.