Add link to the scenario name in the dropup callout
This allows user to click on the failing step in the dropup callout
which takes them directly to the failure point in the feature report.
Step object was refactored to include the line number de-serialized
from the JSON data file which is in-turn used along with the index of
the scenario to create a composite id that is then used to locate the
failing step element on the feature report page. Unit and integration
tests were also modified/expanded to include: testing of the contents
of the callout contained within the failure dropup on the feature
overview page (scenario name, step name/text & url, and error message),
the element id attributes of the scenario steps on the feature overview
page, and the various changes made to the Cucumber object models
(Step, Feature, etc.)