Bump oss-parent to last ever version '9', add relativePath so it forces look to m2 repo (#1126)
* [pom] Add missing relativePath to parent and add comment as its end of life

should use your own or define the values.  That one is so out of date and hasn't been supported in years, was officially deprecated.

* [pom] Bump oss-parent to 9 - the last to ever be released

As noted, the contents should be brought into project as sonatype does not promote this any more but at least be up to date for now.