[enum] Handle enums appropriately
An impossible default to use a switch here is not remotely valuable as its impossible to reach without completely replacing the enum as was done in the tests to solve for non existent default throws.  That hack lights up the IDE's as the class is entirely duplicated.  After enough stack overflows on this then finally reading a write up from the original creator of Jmockit that it was a wrong solution, just wrote it correctly.

In other words, there are 2 enums.  Its either one or the other.  It cannot be something else.  Therefore, it doesn't need a switch statement.  It should only be an 'if' statement.  If it later becomes more, it could change but not seeing why now or why a about 10 year old hack suggested then contested was used here.  Personally I prefer my IDE to be free of errors ;)